Fixes #4716 Add more statistics for events
